fix(reality-server): widen portal_saved_searches.match_count to bigint (#2815)
Follow-up to PR #2812. RealityPortalRepository::mark_saved_search_matched
advances the alert watermark with `match_count = match_count + N`, and
`portal_saved_searches.match_count` was `int4`. Since PR #2812 made the
enqueue + advance atomic (correct), a failed advance now rolls the whole
run back — which exposes a latent overflow edge: a long-lived saved
search accumulating `match_count` near i32::MAX would have every
subsequent advance abort on `integer` overflow, wedging the search and
silently stopping alert delivery forever.

Widen the column to `bigint` (migration 00232), lift the Rust model /
repo types from i32 -> i64, and lock the behavior in with a new
regression test (`watermark_advances_past_i32_max_boundary`) that primes
match_count past i32::MAX and asserts the watermark still moves. The
pre-existing rollback-on-overflow guard is preserved by climbing to
i64::MAX to force the bigint overflow — same contract, new ceiling.

Closes #2814

+++ backend/crates/db/src/repositories/reality_portal/searches.rs
@@ -133,6 +133,12 @@ impl RealityPortalRepository {
 
     /// Advance a saved search's match watermark (`last_matched_at = now()`) and
     /// add `new_matches` to its running `match_count`.
+    ///
+    /// `match_count` is stored as `bigint` (i64) — see migration 00232 and
+    /// #2814. The pre-widening `int4` column could overflow near `i32::MAX`
+    /// and wedge the search forever once PR #2812 made the advance atomic
+    /// (a failed advance rolls the enqueue back). `bigint` pushes the ceiling
+    /// far past any realistic per-search match volume.
     pub async fn mark_saved_search_matched<'e, E>(
         &self,
         executor: E,
@@ -152,7 +158,7 @@ impl RealityPortalRepository {
             "#,
         )
         .bind(id)
-        .bind(new_matches as i32)
+        .bind(new_matches)
         .execute(executor)
         .await?;
         Ok(())
